Move most of RefreshTokenAuthorizer logic into abstract superclass RenewingAuthorizer
